Refactor to split index classes between multiple files
