+++ src/common/plug_indexes/effect.py
@@ -1,0 +1,25 @@
+from .fl_index import PluginIndex
+
+
+class EffectIndex(PluginIndex):
+    def __init__(self, index: int, slotIndex: int) -> None:
+        self.__index = index
+        self.__slot = slotIndex
+
+    def __repr__(self) -> str:
+        return \
+            f"EffectIndex({self.__index}, {self.__slot}, {self.getName()!r})"
+
+    @property
+    def index(self) -> int:
+        """
+        The mixer track that contains the plugin.
+        """
+        return self.__index
+
+    @property
+    def slotIndex(self) -> int:
+        """
+        The mixer slot that contains the plugin.
+        """
+        return self.__slot
