+++ src/plugs/windows/channel_rack/omni.py
@@ -13,7 +13,7 @@ more details.
 from typing import Any
 import channels
 from common.types.color import Color
-from common.plug_indexes import UnsafeIndex
+from common.plug_indexes.fl_index import UnsafeIndex
 from control_surfaces import ControlShadowEvent
 from control_surfaces import (
     DrumPad,
