+++ src/plugs/windows/playlist.py
@@ -19,7 +19,7 @@ import general
 from common import getContext
 from common.types import Color
 from common.extension_manager import ExtensionManager
-from common.plug_indexes import UnsafeIndex
+from common.plug_indexes.fl_index import UnsafeIndex
 from common.util.api_fixes import getFirstPlaylistSelection
 from control_surfaces import consts
 from control_surfaces import ControlShadowEvent
